fix(clippy): clean up 11 lints exposed by --all-features build
`cargo clippy --all-targets --all-features` lit up after libfuse-dev
landed on the ubuntu runner — clippy could finally compile the
fuse-gated paths plus everything else. The lints were a backlog from
recent landings (ext htree, JBD2 multi-descriptor, repack, fuse
adapter); CI had been failing earlier on macOS so nobody saw them.

Library / binary:
- `ext::htree::name_in_hash_buf`: replace `for i in 0..len` with
  `msg.iter().take(len).enumerate()` (needless_range_loop).
- `ext::htree::make_dx_root_block`: `#[allow(clippy::too_many_arguments)]`
  — it mirrors the on-disk `dx_root` shape, splitting it would just
  shuffle state.
- `ext::jbd2::stage_metadata_transaction_at`: collapse the
  `is_empty() ⇒ 1 / ≤ first_cap ⇒ 1` cases into one branch
  (if_same_then_else).
- `ext::Ext::populate_depth1_leaves`: drop redundant `.into_iter()`
  in `.zip(...)`.
- `ext::Ext::patch_extent_block`: `(bs / 4) as u32` → `bs / 4` (bs
  is already `u32`).
- `ext::Ext::rename` doc: rework the `(probe + remove + rename)`
  bullet — the `+ rename)` was being parsed as a markdown list
  continuation.
- `fuse_adapter::attr_for`: drop the `u16 → u16` self-cast.
- `repack::Progress::tick`: `is_multiple_of(500)` is 1.87-stable, we
  pin MSRV 1.85 — replace with `% 500 == 0`.
- `repack::copy_ext_dir_at` + `fstool::main::copy_ext_dir_at` +
  `create_via_factory`: `#[allow(clippy::too_many_arguments)]` —
  recursive walker / generic dispatcher, each arg is a distinct knob.

Tests:
- `tests/ext2_external.rs`: drop `inode.size as u32` self-cast.
- `tests/ext4_external.rs`: replace `drop(jb_phys)` (closure isn't
  `Drop`) with a scoped block that lets the borrow expire naturally.

722 lib tests + every integration suite pass locally with
`RUSTFLAGS=-D warnings cargo clippy --all-targets --all-features`
clean.

+++ src/fuse_adapter.rs
@@ -115,7 +115,7 @@ impl FstoolFs {
             ctime: ts(inode.ctime),
             crtime: ts(inode.mtime),
             kind,
-            perm: (inode.mode & 0o7777) as u16,
+            perm: inode.mode & 0o7777,
             nlink: inode.links_count as u32,
             uid: inode.uid as u32,
             gid: inode.gid as u32,
